requirements.txt
fastapi>=0.115.0
slowapi>=0.1.9
uvicorn[standard]>=0.27.0
pydantic>=2.7.4
httpx>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
moviepy>=1.0.3
pandas>=2.0.0

# Authentication dependencies
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
fastapi-mail>=1.4.1
python-multipart>=0.0.12
cloudinary>=1.36.0
runwayml>=0.1.0
ffmpeg-python
timezonefinder
pytz
